ref: 지도페이지 Map 컴포넌트 분리
