--- one-n/src/pages/MapsPage/MapsPage.module.scss
@@ -1,6 +1,0 @@
-.MapsPage {
-    .map {
-        width: 100%;
-        height: calc(100vh - 6.25rem);
-    }
-}
