+++ one-n/src/pages/MapsPage/sub-components/Map/Map.module.scss
@@ -1,0 +1,4 @@
+.Map {
+    width: 100%;
+    height: calc(100vh - 6.25rem);
+}
